+++ ignition/modules/IsoniaProtocolV01.ts
@@ -5,9 +5,10 @@ export default buildModule("IsoniaProtocolV01Module", (m) => {
 
   const govCore = m.contract("GovCore");
   const demoTarget = m.contract("DemoTarget", [deployer]);
+  const demoVotesToken = m.contract("IsoDemoVotesToken", [deployer]);
   const govProposals = m.contract("GovProposals", [govCore, demoTarget]);
 
   m.call(demoTarget, "setGovProposals", [govProposals]);
 
-  return { govCore, govProposals, demoTarget };
+  return { govCore, govProposals, demoTarget, demoVotesToken };
 });